image: separate imagewriter to own struct
